chore: retire tick-era tooling
