+++ Causal_Web/engine/logging/__init__.py
@@ -1,0 +1,5 @@
+"""Minimal logging helpers for the event engine."""
+
+from .logger import log_record
+
+__all__ = ["log_record"]
