+++ Causal_Web/engine/services/node_services.py
@@ -14,15 +14,12 @@ from dataclasses import dataclass
 from typing import Any, Optional, List, Dict, Set, TYPE_CHECKING
 from collections import defaultdict, deque
 import threading
-from concurrent.futures import ThreadPoolExecutor
 import numpy as np
 
 from ...config import Config
 from ..logging.logger import log_json
 from ..models.tick import Tick, GLOBAL_TICK_POOL
 from ..models.node import Node, NodeType, Edge
-from ..quantum.tensors import propagate_chain
-from ..backend.cupy_kernels import complex_weighted_sum
 
 HADAMARD = np.array([[1, 1], [1, -1]], dtype=np.complex128) / np.sqrt(2)
 
@@ -283,308 +280,14 @@ class NodeTickService:
 
     # ------------------------------------------------------------------
     def _propagate_edges(self, tick: Tick) -> None:
-        if self.node.cnot_source:
-            edges = self.graph.get_edges_from(self.node.id)
-            if len(edges) >= 2:
-                e1, e2 = edges[:2]
-                if not getattr(e1, "epsilon", False) or not getattr(
-                    e2, "epsilon", False
-                ):
-                    e1.epsilon = True
-                    e2.epsilon = True
-                    pair_id = str(uuid.uuid4())
-                    e1.partner = e2
-                    e2.partner = e1
-                    e1.partner_id = pair_id
-                    e2.partner_id = pair_id
-        EdgePropagationService(
-            node=self.node,
-            tick_time=self.tick_time,
-            phase=self.phase,
-            origin=self.origin,
-            graph=self.graph,
-            tick=tick,
-        ).propagate()
+        """Propagate edges for a fired node.
 
-
-@dataclass
-class EdgePropagationService:
-    """Handle tick propagation across outgoing edges.
-
-    Propagation runs in parallel across edges using a thread pool.
-    """
-
-    node: Node
-    tick_time: int
-    phase: float
-    origin: str
-    graph: Any
-    tick: Tick
-
-    def propagate(self) -> None:
-        """Propagate the tick across all outgoing edges.
-
-        Phases and attenuations are batched into arrays and multiplied via
-        :func:`complex_weighted_sum`, which uses CuPy when available.  The
-        resulting weights are then distributed to individual edge handlers.
+        The v2 engine processes events directly, so edge propagation from the
+        legacy tick engine is no longer required. This method now acts as a
+        placeholder and performs no work.
         """
 
-        self._log_recursion()
-        from ..tick_engine import kappa
-
-        edges = self.node._fanout_edges(self.graph)
-        if not edges:
-            return
-
-        if self.node.node_type == NodeType.DECOHERENT:
-            base_phase = (
-                self.node.locked_phase if self.node.locked_phase is not None else 0.0
-            )
-            shifted = np.array([base_phase + e.phase_shift + e.A_phase for e in edges])
-        else:
-            shifted = np.array([self._shift_phase(e) for e in edges])
-        phis = np.exp(1j * shifted)
-        attenuations = np.array([e.attenuation for e in edges])
-        weights = complex_weighted_sum(phis, attenuations)
-
-        def worker(args: tuple[Edge, float, complex]) -> None:
-            edge, s, w = args
-            self._propagate_edge(edge, s, w, kappa)
-
-        with ThreadPoolExecutor(
-            max_workers=getattr(Config, "thread_count", None)
-        ) as ex:
-            ex.map(worker, zip(edges, shifted, weights))
-
-    # ------------------------------------------------------------------
-    def _log_recursion(self) -> None:
-        if self.origin != "self" and any(
-            e.target == self.origin for e in self.graph.get_edges_from(self.node.id)
-        ):
-            log_json(
-                "event",
-                "refraction_log",
-                {"recursion_from": self.origin, "node": self.node.id},
-                tick=self.tick_time,
-            )
-
-    # ------------------------------------------------------------------
-    def _propagate_edge(
-        self, edge: Edge, shifted: float, weighted_phi: complex, kappa: float
-    ) -> None:
-        chain = self._collect_chain(edge)
-        if len(chain) > 4:
-            self._propagate_chain_mps(chain, kappa)
-            return
-        target = self.graph.get_node(edge.target)
-        delay = edge.adjusted_delay(
-            self.node.law_wave_frequency,
-            target.law_wave_frequency,
-            kappa,
-            graph=self.graph,
-        )
-        from ..fields.density import get_field
-
-        rho = get_field().get(edge)
-        delay = max(1.0, delay * (1 + kappa * rho))
-        if self.node.node_type == NodeType.DECOHERENT:
-            psi_contrib = self.node.probabilities * edge.attenuation
-        else:
-            source_psi = self.node.psi
-            if edge.u_id == 1:
-                psi_src = HADAMARD @ source_psi
-            else:
-                psi_src = source_psi
-            psi_contrib = weighted_phi * psi_src
-
-        target.psi += psi_contrib
-        get_field().deposit(edge, psi_contrib)
-        self._log_propagation(target, delay, shifted)
-        new_delay = self.tick.cumulative_delay + delay
-        max_delay = getattr(Config, "max_cumulative_delay", 0)
-        if max_delay and new_delay > max_delay:
-            if getattr(Config, "log_tick_drops", True):
-                target._log_tick_drop(
-                    self.tick_time,
-                    "event_horizon",
-                    tick_id=self.tick.trace_id,
-                    source=self.node.id,
-                    target=target.id,
-                    cumulative_delay=new_delay,
-                    coherence=target.compute_coherence_level(self.tick_time),
-                )
-            return
-        if self._handle_refraction(target, delay, shifted, kappa):
-            return
-        new_tick = GLOBAL_TICK_POOL.acquire()
-        new_tick.origin = self.tick.origin
-        new_tick.time = self.tick.time
-        new_tick.amplitude = self.tick.amplitude * edge.attenuation
-        new_tick.phase = shifted
-        new_tick.layer = self.tick.layer
-        new_tick.trace_id = self.tick.trace_id
-        new_tick.generation_tick = self.tick.generation_tick
-        new_tick.cumulative_delay = new_delay
-        new_tick.entangled_id = self.tick.entangled_id
-        target.schedule_tick(
-            self.tick_time + delay,
-            shifted,
-            origin=self.node.id,
-            created_tick=self.tick_time,
-            amplitude=new_tick.amplitude,
-            tick_id=new_tick.trace_id,
-            cumulative_delay=new_delay,
-            entangled_id=new_tick.entangled_id,
-            graph=self.graph,
-        )
-        GLOBAL_TICK_POOL.release(new_tick)
-
-    # ------------------------------------------------------------------
-    def _collect_chain(self, edge: Edge) -> List[Edge]:
-        chain = [edge]
-        current = self.graph.get_node(edge.target)
-        visited = {edge.source, edge.target}
-        while True:
-            edges = self.graph.get_edges_from(current.id)
-            if len(edges) != 1:
-                break
-            nxt = edges[0]
-            if nxt.target in visited:
-                break
-            chain.append(nxt)
-            visited.add(nxt.target)
-            current = self.graph.get_node(nxt.target)
-        return chain
-
-    # ------------------------------------------------------------------
-    def _propagate_chain_mps(self, chain: List[Edge], kappa: float) -> None:
-        target = self.graph.get_node(chain[-1].target)
-        total_delay = 0.0
-        unitaries: List[np.ndarray] = []
-        attenuation = 1.0
-        shifted = self.phase
-        source_freq = self.node.law_wave_frequency
-        current_freq = source_freq
-        from ..fields.density import get_field
-
-        field = get_field()
-        for e in chain:
-            tgt = self.graph.get_node(e.target)
-            delay = e.adjusted_delay(
-                current_freq,
-                tgt.law_wave_frequency,
-                kappa,
-                graph=self.graph,
-            )
-            rho = field.get(e)
-            delay = delay * (1 + kappa * rho)
-            total_delay += delay
-            current_freq = tgt.law_wave_frequency
-            unitaries.append(
-                HADAMARD if e.u_id == 1 else np.eye(2, dtype=np.complex128)
-            )
-            attenuation *= e.attenuation
-            shifted += e.phase_shift + e.A_phase
-        ei_phi = np.exp(1j * shifted)
-        # TODO: Replace propagate_chain with CuPy kernel
-        psi_contrib, _ = propagate_chain(
-            unitaries, self.node.psi, chi_max=getattr(Config, "chi_max", 16)
-        )
-        psi_contrib = ei_phi * psi_contrib.reshape(-1) * attenuation
-        target.psi += psi_contrib
-        for e in chain:
-            field.deposit(e, psi_contrib)
-        self._log_propagation(target, total_delay, shifted)
-        new_delay = self.tick.cumulative_delay + total_delay
-        max_delay = getattr(Config, "max_cumulative_delay", 0)
-        if max_delay and new_delay > max_delay:
-            if getattr(Config, "log_tick_drops", True):
-                target._log_tick_drop(
-                    self.tick_time,
-                    "event_horizon",
-                    tick_id=self.tick.trace_id,
-                    source=self.node.id,
-                    target=target.id,
-                    cumulative_delay=new_delay,
-                    coherence=target.compute_coherence_level(self.tick_time),
-                )
-            return
-        if self._handle_refraction(target, total_delay, shifted, kappa):
-            return
-        new_tick = GLOBAL_TICK_POOL.acquire()
-        new_tick.origin = self.tick.origin
-        new_tick.time = self.tick.time
-        new_tick.amplitude = self.tick.amplitude * attenuation
-        new_tick.phase = shifted
-        new_tick.layer = self.tick.layer
-        new_tick.trace_id = self.tick.trace_id
-        new_tick.generation_tick = self.tick.generation_tick
-        new_tick.cumulative_delay = new_delay
-        new_tick.entangled_id = self.tick.entangled_id
-        target.schedule_tick(
-            self.tick_time + total_delay,
-            shifted,
-            origin=self.node.id,
-            created_tick=self.tick_time,
-            amplitude=new_tick.amplitude,
-            tick_id=new_tick.trace_id,
-            cumulative_delay=new_delay,
-            entangled_id=new_tick.entangled_id,
-            graph=self.graph,
-        )
-        GLOBAL_TICK_POOL.release(new_tick)
-
-    # ------------------------------------------------------------------
-    def _shift_phase(self, edge: Edge) -> float:
-        return self.phase + edge.phase_shift + edge.A_phase
-
-    # ------------------------------------------------------------------
-    def _log_propagation(self, target: Node, delay: float, shifted: float) -> None:
-        log_json(
-            "event",
-            "tick_propagation_log",
-            {
-                "source": self.node.id,
-                "target": target.id,
-                "arrival_time": self.tick_time + delay,
-                "phase": shifted,
-            },
-            tick=self.tick_time,
-        )
-
-    # ------------------------------------------------------------------
-    def _handle_refraction(
-        self, target: Node, delay: float, shifted: float, kappa: float
-    ) -> bool:
-        if target.node_type != NodeType.DECOHERENT:
-            return False
-        alts = self.graph.get_edges_from(target.id)
-        if not alts:
-            return False
-        alt = alts[0]
-        alt_tgt = self.graph.get_node(alt.target)
-        alt_delay = alt.adjusted_delay(
-            target.law_wave_frequency,
-            alt_tgt.law_wave_frequency,
-            kappa,
-            graph=self.graph,
-        )
-        alt_tgt.schedule_tick(
-            self.tick_time + delay + alt_delay,
-            shifted,
-            origin=self.node.id,
-            created_tick=self.tick_time,
-            entangled_id=self.tick.entangled_id,
-            graph=self.graph,
-        )
-        target.node_type = NodeType.REFRACTIVE
-        log_json(
-            "event",
-            "refraction_log",
-            {"from": self.node.id, "via": target.id, "to": alt_tgt.id},
-            tick=self.tick_time,
-        )
-        return True
+        return None
 
 
 @dataclass
