+++ tests_legacy/test_quantum_lite.py
